requirements.txt
psutil
python-socketio[client]
flask
PyYAML